:recycle: HomeViewModel에서 링크 저장 관련 로직 분리
